Refactor `spec.go`, moving some things into their own files
